Delay pinger on stream bind
Given the serial state machine implemented by lib-gst-meet, it was
expecting an answer to the extdisco:2 iq but got a pong, and wrongly
assumed no STUN/TURN candidates were exposed by the server.

A much better solution would be to match on the iq id as well, but this
will do as a workaround.